MPT-17825: Split excel_file_handler_mixins into per-mixin package
The plural filename was already hinting at this layout. Extract each
mixin into its own module under `cli/core/handlers/excel_mixins/`:

- `workbook.py` — `ExcelWorkbookMixin`
- `validation.py` — `ExcelValidationMixin`
- `read.py` — `ExcelReadMixin` (plus the two protected helpers
  `_build_dynamic_column_map` and `_column_matches`)
- `write.py` — `ExcelWriteMixin`
- `worksheet.py` — `ExcelWorksheetMixin`
- `composites.py` — `ExcelAccessMixin`, `ExcelSheetMixin`
- `types.py` — `SheetData`, `SheetDataGenerator`, `ColumnPatterns`

The package `__init__.py` only re-exports the composite mixins
(`ExcelAccessMixin`, `ExcelSheetMixin`) that `excel_file_handler.py`
consumes. Atomic mixins and type aliases remain implementation
details: callers must import them from their specific submodule
(no external code currently does so).

Other touches inherited from the split:

- All type aliases (including the ones only used by a single submodule)
  live in `types.py` so that adding new ones doesn't risk silent
  duplication across modules.
- Class attribute declarations follow the project convention:
  protected first (alphabetical), then public (alphabetical).
- `_build_dynamic_column_map` is no longer `@classmethod` because it
  never used `cls`. It is a regular protected method, alongside the
  new `_column_matches` helper that keeps the boolean condition
  readable without tripping WPS221.

Importers updated:

- `cli/core/handlers/excel_file_handler.py` now imports the
  composites from `cli.core.handlers.excel_mixins`.
- `tests/cli/core/handlers/test_excel_file_handler.py` patches
  `cli.core.handlers.excel_mixins.workbook.load_workbook` (the new
  location of the `load_workbook` import).

No behavior changes. The two `# noqa: WPS214` on `ExcelReadMixin`
and `ExcelWriteMixin` stay (WPS214 is per-class, not per-file, and
splitting modules does not move the method count below threshold).

Validation: ruff format/check, flake8, mypy, pytest (364 passed,
coverage 98%) all green.
